+++ src/Autocomplete/src/Controller/EntityAutocompleteController.php
@@ -15,6 +15,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\UX\Autocomplete\AutocompleteResultsExecutor;
 use Symfony\UX\Autocomplete\AutocompleterRegistry;
 
@@ -27,7 +28,8 @@ final class EntityAutocompleteController
 {
     public function __construct(
         private AutocompleterRegistry $autocompleteFieldRegistry,
-        private AutocompleteResultsExecutor $autocompleteResultsExecutor
+        private AutocompleteResultsExecutor $autocompleteResultsExecutor,
+        private UrlGeneratorInterface $urlGenerator,
     ) {
     }
 
@@ -38,10 +40,20 @@ final class EntityAutocompleteController
             throw new NotFoundHttpException(sprintf('No autocompleter found for "%s". Available autocompleters are: (%s)', $alias, implode(', ', $this->autocompleteFieldRegistry->getAutocompleterNames())));
         }
 
-        $results = $this->autocompleteResultsExecutor->fetchResults($autocompleter, $request->query->get('query', ''));
+        $page = $request->query->getInt('page', 1);
+        $nextPage = null;
+
+        $data = $this->autocompleteResultsExecutor->fetchResults($autocompleter, $request->query->get('query', ''), $page);
+
+        if ($data->hasNextPage) {
+            $parameters = array_merge($request->attributes->all('_route_params'), $request->query->all(), ['page' => $page + 1]);
+
+            $nextPage = $this->urlGenerator->generate($request->attributes->get('_route'), $parameters, UrlGeneratorInterface::ABSOLUTE_URL);
+        }
 
         return new JsonResponse([
-            'results' => $results,
+            'results' => $data->results,
+            'next_page' => $nextPage,
         ]);
     }
 }