+++ src/Autocomplete/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2.5.0
+
+-   Automatic pagination support added: if the query would return more results
+    than your limit, when the user scrolls to the bottom of the options, it will
+    make a second Ajax call to load more.
+
 ## 2.4.0
 
 -   Support added for setting the required minimum search query length (defaults to 3) (#492) - @daFish